Execute git commands in a subshell
When git-num is executed with `git num` (as opposed to calling
`path/to/git-num` directly), a SIGPIPE from a child process (e.g.
`less`) causes a noisy failure: "git-num died of signal 13". This is not
an issue if commands are executed in a subshell rather than as a
replacement for the git-num process (via exec()).

Fixes #5